requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==4.2.20
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_metadata==8.5.0
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
zipp==3.20.2
